+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,17 @@
 # sit-onyx
 
+## 1.0.0-beta.118
+
+### Minor Changes
+
+- 9ddcd26: feat: implement `OnyxCard` component
+
+## 1.0.0-beta.117
+
+### Minor Changes
+
+- 80bdfb2: fix(OnyxBottomBar): Adjust styles to reflect the grid. Added screenshot tests and a bottom bar in the demo app.
+
 ## 1.0.0-beta.116
 
 ### Minor Changes